+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * [Fixed] randomize_call bug
 * [Fixed] parameter override by class type bug
 * [Fixed] hierarchical this bug
+* [Fixed] hierarchical delay value bug
+* [Fixed] const class new bug
 
 ## [v0.3.6](https://github.com/dalance/sv-parser/compare/v0.3.5...v0.3.6) - 2019-11-05